Initial tests for the macro extractor
