+++ MacroExtractor/RepairSequence.cs
@@ -10,11 +10,13 @@ namespace MacroExtractor
 {
     public class RepairSequence
     {
+        public GroundedAction MetaAction { get; set; }
         public ActionDecl Macro { get; set; }
         public ActionPlan Replacement { get; set; }
 
-        public RepairSequence(ActionDecl macro, ActionPlan replacement)
+        public RepairSequence(GroundedAction metaAction, ActionDecl macro, ActionPlan replacement)
         {
+            MetaAction = metaAction;
             Macro = macro;
             Replacement = replacement;
         }
