+++ P9.sln
@@ -38,10 +38,20 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PreBuilder", "PreBuilder\PreBuilder.csproj", "{A7EAD3AB-2CA4-4485-8DC5-BDC5A2FE9884}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TestingSets", "TestingSets", "{8DA6DBD4-939F-4D75-9A59-98010AA30B59}"
+	ProjectSection(SolutionItems) = preProject
+		TestingSets\blocksworld_p01-p19.zip = TestingSets\blocksworld_p01-p19.zip
+		TestingSets\ferry_p01-p19.zip = TestingSets\ferry_p01-p19.zip
+		TestingSets\floortile_p01-p19.zip = TestingSets\floortile_p01-p19.zip
+		TestingSets\miconic_p01-p19.zip = TestingSets\miconic_p01-p19.zip
+		TestingSets\rovers_p01-p19.zip = TestingSets\rovers_p01-p19.zip
+		TestingSets\satellite_p01-p19.zip = TestingSets\satellite_p01-p19.zip
+	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{8B977624-590C-429F-AF65-6C2B00C75AF1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MetaActionGenerator.Tests", "Tests\MetaActionGenerator.Tests\MetaActionGenerator.Tests.csproj", "{5C5289EA-288A-4728-91E6-DC001AD21A7C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MetaActionGenerator.Tests", "Tests\MetaActionGenerator.Tests\MetaActionGenerator.Tests.csproj", "{5C5289EA-288A-4728-91E6-DC001AD21A7C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MacroExtractor.Tests", "Tests\MacroExtractor.Tests\MacroExtractor.Tests.csproj", "{3CD2C382-CD00-46E5-807F-BEE118867793}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -85,12 +95,17 @@ Global
 		{5C5289EA-288A-4728-91E6-DC001AD21A7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5C5289EA-288A-4728-91E6-DC001AD21A7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5C5289EA-288A-4728-91E6-DC001AD21A7C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3CD2C382-CD00-46E5-807F-BEE118867793}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3CD2C382-CD00-46E5-807F-BEE118867793}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3CD2C382-CD00-46E5-807F-BEE118867793}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3CD2C382-CD00-46E5-807F-BEE118867793}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{5C5289EA-288A-4728-91E6-DC001AD21A7C} = {8B977624-590C-429F-AF65-6C2B00C75AF1}
+		{3CD2C382-CD00-46E5-807F-BEE118867793} = {8B977624-590C-429F-AF65-6C2B00C75AF1}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3B73CE61-35D1-4280-9F91-DD54F6DAFCD2}
